Rework the API so that /Safeguard - gets the current configuration, puts a new configuration or deletes the ConfigurationName /Logon and /Logoff - logon and logoff the user given a token in the authorize header /ClientCertificate - Gets the current certificate, posts a new signed certificate or pfx in base64, deletes the client certificate /CSR - Gets a new CSR to be signed and posted back to the devops service
When the devops service is configure using PUT /Safeguard?configure=true, a new A2A user is created, a new A2A registration is created and the client certificate is added to Safeguard trusted certificates.  If any of these objects already exist in the target safeguard appliance, the objects are updated with any new values that are required to connect the devops service.

TODO:
- Map the accounts to the A2A RegistrationsTableName
- Map the plugins to the accounts.