Google Cloud Secret Manager Plug-in
Added new plug-in for Google Cloud Secret Manager. Supports reverse flow.

All secrets in Google Cloud are simply Base64 encoded values, there is no distinction for a "credential type". Therefore, this is only configured to support the `CredentialType.Password` for now.

Added plug-in project template project. Both as a pre-zipped ready to use file, as well as the raw template files themselves that can be modified and zipped up.